[REF] merge louve_custom_pricetag into louve_custom_product
