--- louve_addons/louve_custom_pricetag/README.rst
@@ -1,27 +1,0 @@
-.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
-    :alt: License: AGPL-3
-
-==============================
-La Louve - Custom Product Pricetags
-==============================
-
-This module adds some fields on the product that will be displayed on the 
-product pricetag
-
-Installation
-============
-
-Normal installation.
-
-Configuration
-=============
-
-No configuration
-
-Credits
-=======
-
-Contributors
-------------
-
-* Julien Weste <julien.weste@akretion.com.br>
