--- louve_addons/louve_custom_pricetag/data/report_paperformat.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record id="paperformat_pricetag_vegetables" model="report.paperformat">
-        <field name="name">Vegetables pricetag paper</field>
-        <field name="default" eval="True" />
-        <field name="format">A4</field>
-        <field name="page_height">0</field>
-        <field name="page_width">0</field>
-        <field name="orientation">Landscape</field>
-        <field name="margin_top">8</field>
-        <field name="margin_bottom">8</field>
-        <field name="margin_left">8</field>
-        <field name="margin_right">8</field>
-        <field name="header_line" eval="False" />
-        <field name="header_spacing">0</field>
-        <field name="dpi">90</field>
-    </record>
-</odoo>
