--- louve_addons/louve_custom_pricetag/i18n/fr.po
@@ -1,78 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-#   * louve_custom_pricetag
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 9.0c\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2016-10-10 16:57+0000\n"
-"PO-Revision-Date: 2016-10-10 16:57+0000\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: louve_custom_pricetag
-#: model:ir.model.fields,field_description:louve_custom_pricetag.field_product_template_farming_method
-msgid "Farming Method"
-msgstr "Méthode de production"
-
-#. module: louve_custom_pricetag
-#: model:ir.model.fields,help:louve_custom_pricetag.field_product_template_rack_instruction
-msgid "For example, the number of packages that\n"
-"        should be stored on the rack"
-msgstr "Par exemple, le nombre de colis qui\n"
-"        doivent être rayonnés"
-
-#. module: louve_custom_pricetag
-#: model:ir.model.fields,field_description:louve_custom_pricetag.field_product_template_pricetag_louveinfos
-msgid "La Louve custom fields"
-msgstr "Champs personnalisés pour La Louve"
-
-#. module: louve_custom_pricetag
-#: model:ir.model.fields,field_description:louve_custom_pricetag.field_product_template_pricetag_rackinfos
-msgid "La Louve rack fields"
-msgstr "Champs de stockage pour La Louve"
-
-#. module: louve_custom_pricetag
-#: model:ir.model.fields,field_description:louve_custom_pricetag.field_product_template_rack_number_of_packages
-msgid "Number of packages on the rack"
-msgstr "Nombre de colis à rayonner"
-
-#. module: louve_custom_pricetag
-#: model:ir.model.fields,help:louve_custom_pricetag.field_product_template_farming_method
-msgid "Organic Label"
-msgstr "Label bio"
-
-#. module: louve_custom_pricetag
-#: model:ir.model.fields,field_description:louve_custom_pricetag.field_product_template_other_information
-msgid "Other Information"
-msgstr "Autres Informations"
-
-#. module: louve_custom_pricetag
-#: model:ir.model,name:louve_custom_pricetag.model_product_template
-msgid "Product Template"
-msgstr "Modèle d'article"
-
-#. module: louve_custom_pricetag
-#: model:ir.model.fields,field_description:louve_custom_pricetag.field_product_template_rack_instruction
-msgid "Rack Instruction"
-msgstr "Instructions rayonnage"
-
-#. module: louve_custom_pricetag
-#: model:ir.model.fields,field_description:louve_custom_pricetag.field_product_template_rack_location
-msgid "Rack Location"
-msgstr "Emplacement rayon"
-
-#. module: louve_custom_pricetag
-#: model:ir.ui.view,arch_db:louve_custom_pricetag.view_product_template_form_sticker
-msgid "Sticker Information"
-msgstr "Information étiquette"
-
-#. module: louve_custom_pricetag
-#: model:ir.model.fields,help:louve_custom_pricetag.field_product_template_rack_location
-msgid "The name or place of the rack"
-msgstr "Le nom ou emplacement du rayon"
