--- louve_addons/louve_custom_pricetag/model/product_category_print.py
@@ -1,12 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright (C) 2016-Today: La Louve (<http://www.lalouve.net/>)
-# @author: Sylvain LE GAL (https://twitter.com/legalsylvain)
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-from openerp import fields, models
-
-
-class ProductCategoryPrint(models.Model):
-    _inherit = 'product.category.print'
-
-    is_default = fields.Boolean(string='Is default')
