--- louve_addons/louve_custom_pricetag/views/view_product_category_print.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0"?>
-<!--
-    Copyright (C) 2016-Today: La Louve (<http://www.lalouve.net/>)
-    @author: Sylvain LE GAL (https://twitter.com/legalsylvain)
-    License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
--->
-
-<odoo>
-    <record model="ir.ui.view" id="view_product_product_form_print">
-        <field name="model">product.category.print</field>
-        <field name="inherit_id" ref="louve_custom_product.view_product_product_form_print" />
-        <field name="arch" type="xml">
-            <field name="pricetag_model_id" position="after">
-                <field name="is_default" />
-            </field>
-        </field>
-    </record>
-</odoo>
