--- louve_addons/louve_custom_pricetag/views/view_product_template.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0"?>
-<!--
-    Copyright (C) 2016-Today: La Louve (<http://www.lalouve.net/>)
-    @author: Julien Weste
-    License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
--->
-
-<odoo>
-
-    <record id="view_product_template_form_sticker" model="ir.ui.view">
-        <field name="model">product.template</field>
-        <field name="inherit_id" ref="product.product_template_form_view"/>
-        <field name="arch" type="xml">
-            <xpath expr="//group[@name='group_standard_price']" position="after">
-                <group string="Sticker Information" name="sticker_info">
-                    <field name="rack_instruction"/>
-                    <field name="rack_location"/>
-                    <field name="rack_number_of_packages"/>
-                    <field name="farming_method"/>
-                    <field name="other_information"/>
-                </group>
-            </xpath>
-        </field>
-    </record>
-
-</odoo>
