+++ louve_addons/louve_custom_product/data/pricetag_model.xml
