+++ louve_addons/louve_custom_product/data/report_paperformat.xml
@@ -20,4 +20,20 @@
         <field name="pricetag_paperformat_id" ref="louve_custom_product.paperformat_pricetag"/>
         <field name="report_model">louve_custom_product.report_pricetag</field>
     </record>
+
+    <record id="paperformat_pricetag_vegetables" model="report.paperformat">
+        <field name="name">Vegetables pricetag paper</field>
+        <field name="default" eval="True" />
+        <field name="format">A4</field>
+        <field name="page_height">0</field>
+        <field name="page_width">0</field>
+        <field name="orientation">Landscape</field>
+        <field name="margin_top">8</field>
+        <field name="margin_bottom">8</field>
+        <field name="margin_left">8</field>
+        <field name="margin_right">8</field>
+        <field name="header_line" eval="False" />
+        <field name="header_spacing">0</field>
+        <field name="dpi">90</field>
+    </record>
 </odoo>
