+++ louve_addons/louve_custom_product/report/report_pricetag.xml
@@ -32,6 +32,8 @@
                         <div class="label_bottom">
                             <div class="product_informations">
                                 <t name="pricetag_origine" t-if="product['product'].pricetag_origin">Origine: <t t-esc="product['product'].pricetag_origin"/><br/></t>
+                                <t t-if="product['product'].pricetag_rackinfos"><t t-esc="product['product'].pricetag_rackinfos"/><br/></t>
+                                <t t-if="product['product'].pricetag_louveinfos"><t t-esc="product['product'].pricetag_louveinfos"/><br/></t>
                             </div>
                         </div>
                     </div>
