+++ louve_addons/louve_custom_product/report/report_pricetag_vegetables.py
