+++ louve_addons/louve_custom_product/static/src/css/pricetag_vegetables.css
